requirements.txt
python-dotenv
langchain
openai
langchain-community
langchain-core
langchain-openai
langchain-chroma 
langgraph
langchain
flask